Update trait.process.R
I think the equations used for the BM and OU processes need to be updated. For the BM process, mvrnorm expects a variance-covariance matrix Sigma, not a correlation matrix so no squaring is required- we should just sample from Sigma*edge.length. For the OU process, the expected mean and variance of an OU process should depend on edge.length (i.e time), so that they decay to the OU optimum and the stationary variance respectively as time goes to infinity. For the same reason I made changes to the multi-peak model. In its current form the OU process model only allows for a single alpha across all traits. Also, the assumption that the OU process can be modelled as a sample from a multivariate normal distribution only holds asymptotically as time goes to infinity. Therefore, you may wish to consider incorporating an OU model from another package that does this exactly rather than using an approximation.